Fix rubocop formatter
Rubocop outputs both the violations it detected and the file contents to
STDOUT. The command tried to pipe this to awk in order to cut off the
first half and only output the formatted file contents.

This did not work as the command wasn't run in a shell and the pipes and
shell constructs were passed as real command line arguments, which
Rubocop confused to be other files to check. The `--stdin` option is not
allowed together with filenames, and so Rubocop exited with an error.

Adding the `--stderr` option to Rubocop makes it output only the file
contents to STDOUT and the violations to STDERR.

Added a new option called `ignore_stderr` that resets STDERR after
running the command so the command doesn't fail because of operator
output.

Now the formatter works correctly.