fix(builtins): Make finding credo cwd smarter
Make the cwd smarter by searching upwards from the buffer, limited to
the project root, and looking for a `.credo.exs` file. While optional,
most projects will have a config file. A conditional can be added to a
user's config to skip the diagnostic if there is no `.credo.exs` file in
the directory.

Combining stdin with `json_raw` allows for more rapid feedback while
still reporting errors to the user, such as incomplete files mid-edit.