fix(diagnostics): Use raw substring for credo
Instead of attempting to decode json and falling back to a substring,
try to find json in the raw output first. This should also hopefully
solve some occasional errors when the output having some values that the
decode doesn't like.

Also tried to get rid of the nested conditionals with some early
returns.