fix: Add support for dotted filetypes
Neovim supports setting multiple filetypes through the use of a dot
separator. From `:h 'filetype'`:

    When a dot appears in the value then this separates two filetype
    names.  Example:
        /* vim: set filetype=c.doxygen : */ ~
    This will use the "c" filetype first, then the "doxygen" filetype.
    This works both for filetype plugins and for syntax files.  More than
    one dot may appear.

By respecting this functionality, users can set filetypes such as
`json.openapi` or `yaml.docker-compose` and get out-of-the-box support
for multiple filetype-specific sources in a single file.

Dotted filetypes continue to be supported as a unique filetype for
backward compatibility and cases where sources should only be made
available at the intersection of filetypes.

For files where a source is explicitly configured to be both enabled and
disabled from different filetypes, I opted to give precedence to the
disabled filetype. It's trivial to swap if that behavior is less
desirable. In practice this might never come up.