Fix mypy builtin not displaying some diagnostics (#873)
The mypy builtin was not displaying some diagnostics due to failing to
parse them properly. The previously used regex pattern assumed that
each line in mypy's output would have a column number and a error code.
This is the case for most lines but not for all of them. For example the
following line contains no column number or error code:

tests/slack_app/conftest.py:10: error: Unused "type: ignore" comment

In this case null-ls would fail to display this error in the editor
because it could not be parsed. This commit fixes that issue by adding a
additional regex pattern that can parse lines in the format of the
example.