fix(rubocop): parse rubocop offenses for better diagnostic output
Some RuboCop offenses, such as Metrics/ClassLength, return an offense
start and end line/column range of the entire class. This means if you
have Neovim underline diagnostics enabled, the whole class will get
underlined, which is a lot of noise and unhelpful.

I looked up how Solargraph, the most popular Ruby LSP which uses
RuboCop, handles this situation. If the end line is different from the
start line, it says the offense ends on the start line at column 0. This
has the effect of underlining just the class declaration instead of the
whole thing.

References:
https://github.com/castwide/solargraph/blob/81bd810e7578040cbdb7650b71abb5c8a10962ab/lib/solargraph/diagnostics/rubocop.rb#L87-L95