fix(builtins): keep ansiblelint in sync (#1240)
Ansiblelint does not accept input from stdin, so either the null-ls
builtin needs to use `methods.internal.DIAGNOSTICS_ON_SAVE` to run
ansiblelint on save or it needs to use `to_temp_file` so the buffer
content is written to a temporary file on which ansiblelint can run.
Otherwise ansiblelint always runs on the previous modification.
Ansiblelint is fast enough so `to_temp_file` works well, but it changes
the filename in the diagnostic. In combination with `multiple_files`
that means that diagnostics do not show up. `multiple_files` seemed
unnecessary anyway so removing it should not cause problems.