fix(mypy): disable multiple files
When a file is passed to mypy, it only checks that file, so the plugin
should not handle filename.

Also, filename returned by mypy is relative to cwd of where mypy is run.
This cwd may be different from the current cwd of neovim, this result of
diagnostic.filename to be unpredictable as neovim toolings will concat cwd to
diagnostic.filename to guess the right buffer.