fix(mypy): Fix wrong filenames in mypy diagnostics
mypy has a wrong configuration, providing erroneous combination of
"on_output" and "cwd" args for generator_factory.

In this configuration cwd often returns the root of the project, which
contains some config such as "pyproject.toml" or "mypy.ini".
The problem is that in the same time "on_output" does not take into
account such non-standard cwd and returns mypy's output as is, which has
relative (!) paths. As a result, null-ls registers buffers, relative to
the root of the project (rather than the actual cwd).

Neovim fails to show such diagnostics (because it has wrong paths) and
some other tools, for example "folke/trouble.nvim" also fail, because
they try to set non-trivial cursor position to a non-existent buffer.

This commit enables showing absolute paths in mypy's config.